Last commit

+++ src/main/java/com/example/demo/service/impl/AppointmentServiceImpl.java
@@ -10,8 +10,9 @@ import com.example.demo.service.mapper.AppointmentMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-import java.util.Optional;
+import java.time.LocalDate;
+import java.time.ZoneId;
+import java.util.*;
 
 @Service
 public class AppointmentServiceImpl implements AppointmentService {
@@ -23,9 +24,25 @@ public class AppointmentServiceImpl implements AppointmentService {
     @Autowired
     DoctorService doctorService;
 
+    @Override
     public Appointment addAppointment(AppointmentCreateDTO dto) {
-        Appointment newAppointment = appointmentRepository.save(appointmentMapper.toAppointment(dto));
-        return newAppointment;
+        List<Doctor> doctorList = doctorService.findByCenterId(dto.centerId);
+        Doctor appointmentDoctor = doctorList.get(0);
+        int minAppointments = findAllByDoctorId(appointmentDoctor.getId()).size();
+        doctorList.remove(appointmentDoctor);
+
+        for(Doctor d : doctorList) {
+            int count = findAllByDoctorId(d.getId()).size();
+            if (count < minAppointments) {
+                appointmentDoctor = d;
+                minAppointments = count;
+            }
+        }
+
+        Appointment newAppointment = appointmentMapper.toAppointment(dto);
+        newAppointment.setDoctor(appointmentDoctor);
+
+        return appointmentRepository.save(newAppointment);
     }
 
     @Override
@@ -35,10 +52,25 @@ public class AppointmentServiceImpl implements AppointmentService {
     }
 
     @Override
-    public void deleteAppointment(Integer id) {
+    public Boolean deleteAppointment(Integer id) {
         Optional<Appointment> appointmentToDelete = appointmentRepository.findById(id);
-        if (appointmentToDelete.isPresent())
-            appointmentRepository.delete(appointmentToDelete.get());
+        Date appointmentDate = appointmentToDelete.get().getDate();
+        if (appointmentToDelete.isPresent()) {
+            Date currentDate = Date.from(LocalDate.now().atStartOfDay(ZoneId.systemDefault()).toInstant());
+            if (appointmentDate.compareTo(currentDate) > 0) {
+                appointmentRepository.delete(appointmentToDelete.get());
+                return true;
+            }
+        }
+        return false;
+    }
+
+    @Override
+    public void deleteAppointmentsByDoctorId(Integer doctorId) {
+        List<Appointment> appointmentsToDelete = appointmentRepository.findAllByDoctorId(doctorId);
+        for(Appointment a : appointmentsToDelete) {
+            deleteAppointment(a.getId());
+        }
     }
 
     @Override
@@ -48,14 +80,30 @@ public class AppointmentServiceImpl implements AppointmentService {
     }
 
     @Override
-    public Appointment confirmAppointment(Integer id, Integer doctorId) {
-        Doctor doctor = doctorService.findById(doctorId);
+    public List<Appointment> findAllByDoctorId(Integer doctorId) {
+        List<Appointment> appointmentsByDoctor = appointmentRepository.findAllByDoctorId(doctorId);
+        return appointmentsByDoctor;
+    }
+
+    @Override
+    public Appointment confirmAppointment(Integer id) {
         Optional<Appointment> confirmedAppointment = appointmentRepository.findById(id);
         if (confirmedAppointment.isPresent()) {
             confirmedAppointment.get().setConfirmed(Boolean.TRUE);
-            confirmedAppointment.get().setDoctor(doctor);
             return appointmentRepository.save(confirmedAppointment.get());
         }
         else return null;
     }
+
+    @Override
+    public List<Appointment> findAllByDoctorAndDate(Integer doctorId, Date date) {
+        List<Appointment> appointments = appointmentRepository.findAllByDoctorIdAndDate(doctorId, date);
+        return appointments;
+    }
+
+    @Override
+    public List<Appointment> findAllByDate(java.sql.Date date) {
+        return appointmentRepository.findAllByDate(date);
+    }
+
 }
